This commit replaces all instances of `endl` with `Qt::endl` across the codebase. The update is in line with the deprecation warnings introduced in Qt 5.14 and ensures compatibility with future versions of Qt. (#124)
Right now, compiling Powertune on more modern version of QT (like the LTS 5.15) generates over 50 warnings.

Reasons for switching to Qt::endl:
1. **Consistency**: `Qt::endl` aligns with Qt naming conventions and integrates better with the Qt framework.
2. **Performance**: Unlike `endl`, which always flushes the output stream, `Qt::endl` provides better control over when the stream is flushed. This can improve performance, especially in scenarios where frequent flushing is unnecessary and can be managed explicitly.
3. **Future-Proofing**: Adhering to the latest standards in Qt ensures easier maintenance and fewer modifications during future upgrades.

The Qt documentation for QTextStream explains that endl is a manipulator that inserts a newline character into the stream and flushes the stream. This automatic flushing can lead to performance overhead when used in loops or high-frequency logging. Switching to Qt::endl allows developers to
 control flushing more explicitly, which can optimize performance by reducing unnecessary flushes.